refactor(server): extract fleet routes into sibling crate
Closes the ADR-0049 follow-up that flagged the fleet HTTP routes as
the natural sibling-crate extraction target once convergio-server
hit the per-crate context-budget cap during F3-5.

Two new crates:

- convergio-server-core - shared HTTP primitives (AppState +
  ApiError enum, From impls for every layer error, the canonical
  IntoResponse mapping). Everything that used crate::error::ApiError
  or crate::app::AppState keeps working via re-exports; behaviour is
  byte-identical.
- convergio-fleet-routes - the 5 fleet route files (fleet,
  fleet_duplicates, fleet_plans, fleet_rot, fleet_doc_drift) plus a
  router() entry point that convergio-server mounts.

Server drops 14150 -> 13075 lines. The 14000 -> 14500 cap bump
introduced during F3-5 is rolled back in check-context-budget.sh.
No route, query parameter, request shape, status code, or error JSON
shape changes - this is a pure structural move.

Verification:
- cargo fmt --all -- --check
- RUSTFLAGS=-Dwarnings cargo clippy --workspace --all-targets
- cargo test --workspace - all green
- bash scripts/check-context-budget.sh - SOFT-WARN (no FAIL)

Tracks: ADR-0049 follow-up row "ADR for extracting fleet_* routes".
