+++ Cargo.toml
@@ -23,6 +23,8 @@ members = [
     "crates/convergio-coherence",
     "crates/convergio-parse-multi",
     "crates/convergio-fleet",
+    "crates/convergio-fleet-routes",
+    "crates/convergio-server-core",
 ]
 resolver = "2"
 
@@ -134,6 +136,8 @@ convergio-cli-session = { path = "crates/convergio-cli-session" }
 convergio-coherence = { path = "crates/convergio-coherence" }
 convergio-parse-multi = { path = "crates/convergio-parse-multi" }
 convergio-fleet = { path = "crates/convergio-fleet" }
+convergio-fleet-routes = { path = "crates/convergio-fleet-routes" }
+convergio-server-core = { path = "crates/convergio-server-core" }
 
 [workspace.lints.rust]
 missing_docs = "warn"
