+++ crates/convergio-fleet-routes/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "convergio-fleet-routes"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+license.workspace = true
+repository.workspace = true
+description = "Fleet HTTP route module (ADR-0038) — extracted from convergio-server."
+readme = "AGENTS.md"
+
+[lints]
+workspace = true
+
+[dependencies]
+convergio-durability = { workspace = true }
+convergio-embed = { workspace = true }
+convergio-fleet = { workspace = true }
+convergio-server-core = { workspace = true }
+convergio-thor = { workspace = true }
+
+axum = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+tokio = { workspace = true }
+tracing = { workspace = true }
