+++ crates/convergio-fleet-routes/src/fleet_plans.rs
@@ -12,13 +12,13 @@
 //! aggregate column on `fleet_plans`. The link insertion is
 //! idempotent on `(fleet_plan_id, repo)`.
 
-use crate::app::AppState;
-use crate::error::ApiError;
 use axum::extract::{Path, Query, State};
 use axum::routing::{get, post};
 use axum::{Json, Router};
 use convergio_durability::{Durability, NewTask};
 use convergio_fleet::{FleetError, FleetPlanRepoLink, FleetPlanView, NewFleetPlan};
+use convergio_server_core::ApiError;
+use convergio_server_core::AppState;
 use convergio_thor::{Thor, Verdict};
 use serde::{Deserialize, Serialize};
 use serde_json::{json, Value};
