+++ crates/convergio-server-core/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "convergio-server-core"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+license.workspace = true
+repository.workspace = true
+description = "Shared HTTP-layer primitives for the Convergio daemon: AppState + ApiError + IntoResponse mapping."
+readme = "AGENTS.md"
+
+[lints]
+workspace = true
+
+[dependencies]
+convergio-bus = { workspace = true }
+convergio-durability = { workspace = true }
+convergio-embed = { workspace = true }
+convergio-fleet = { workspace = true }
+convergio-graph = { workspace = true }
+convergio-lifecycle = { workspace = true }
+convergio-planner = { workspace = true }
+convergio-thor = { workspace = true }
+
+axum = { workspace = true }
+serde_json = { workspace = true }
