+++ crates/convergio-server/AGENTS.md
@@ -19,12 +19,9 @@ The block below is rewritten by `cvg docs regenerate` (ADR-0015) —
 do not edit between the markers.
 
 <!-- BEGIN AUTO:crate_stats -->
-**`convergio-server` stats:** 40 `*.rs` files / 45 public items / 5279 lines (under `src/`).
+**`convergio-server` stats:** 35 `*.rs` files / 40 public items / 4215 lines (under `src/`).
 
 Files approaching the 300-line cap:
 - `src/routes/graph.rs` (295 lines)
-- `src/routes/fleet.rs` (289 lines)
-- `src/routes/fleet_plans.rs` (284 lines)
-- `src/error.rs` (270 lines)
 - `src/routes/audit/append.rs` (254 lines)
 <!-- END AUTO -->
