+++ crates/convergio-server/src/app.rs
@@ -1,46 +1,14 @@
-//! Router assembly + shared state.
+//! Router assembly + shared state re-export.
+//!
+//! [`AppState`] lives in `convergio-server-core` so sibling route
+//! crates can take `State<AppState>` without depending on this
+//! binary crate. This module owns the top-level `Router` assembly
+//! and the middleware layer.
 
 use axum::Router;
-use convergio_bus::Bus;
-use convergio_durability::audit::VerifyReport;
-use convergio_durability::Durability;
-use convergio_embed::{EmbedStore, Embedder};
-use convergio_fleet::{FleetPlanStore, FleetStore};
-use convergio_graph::Store as GraphStore;
-use convergio_lifecycle::Supervisor;
-use std::sync::{Arc, Mutex};
 use tower_http::trace::TraceLayer;
 
-/// Application state injected into every handler.
-#[derive(Clone)]
-pub struct AppState {
-    /// Layer 1 facade.
-    pub durability: Arc<Durability>,
-    /// Layer 2 facade.
-    pub bus: Arc<Bus>,
-    /// Layer 3 facade.
-    pub supervisor: Arc<Supervisor>,
-    /// Tier-3 retrieval store (ADR-0014).
-    pub graph: Arc<GraphStore>,
-    /// Tier-3 semantic embeddings store (ADR-0038, F1).
-    pub embed: Arc<EmbedStore>,
-    /// Embedder used by the daemon for ingest + semantic queries.
-    /// The default at startup is `DeterministicTestEmbedder` (no
-    /// network); set `CONVERGIO_EMBED_MODEL=bge-m3-small-int8` (or
-    /// `multilingual-e5-small`) and build with `--features fastembed`
-    /// to swap in the real ONNX embedder.
-    pub embedder: Arc<dyn Embedder>,
-    /// Fleet repo registry (ADR-0038, F2-6).
-    pub fleet: Arc<FleetStore>,
-    /// Fleet plan store (ADR-0038, F3-2): cross-repo plans with
-    /// per-repo plan links.
-    pub fleet_plans: Arc<FleetPlanStore>,
-    /// Memoised full-chain audit verify result. Keyed by tail `seq`; auto-
-    /// invalidated when a new audit row is appended (tail advances). Shared
-    /// across all clones via `Arc` — one warm call benefits every concurrent
-    /// request. Only applies to the parameter-free `/v1/audit/verify` call.
-    pub audit_verify_cache: Arc<Mutex<Option<(i64, VerifyReport)>>>,
-}
+pub use convergio_server_core::AppState;
 
 /// Build the top-level router. Test harnesses call this directly with
 /// tempdir-backed facades.
@@ -66,8 +34,7 @@ pub fn router(state: AppState) -> Router {
         .merge(crate::routes::workspace::router())
         .merge(crate::routes::graph::router())
         .merge(crate::routes::embed::router())
-        .merge(crate::routes::fleet::router())
-        .merge(crate::routes::fleet_plans::router())
+        .merge(convergio_fleet_routes::router())
         .merge(crate::routes::telemetry::router())
         .merge(crate::routes::api_actions::router())
         .merge(crate::routes::gate_preconditions::router())
