+++ crates/convergio-server/src/routes/mod.rs
@@ -10,11 +10,6 @@ pub mod crdt;
 pub mod dispatch;
 pub mod embed;
 pub mod evidence;
-pub mod fleet;
-pub(crate) mod fleet_doc_drift;
-pub(crate) mod fleet_duplicates;
-pub mod fleet_plans;
-pub(crate) mod fleet_rot;
 pub mod gate_preconditions;
 pub mod graph;
 pub mod health;
