+++ docs/adr/0049-f3-fleet-retrospective.md
@@ -137,7 +137,7 @@ that Roberto runs the live fleet task before declaring F4 ready.
 |------|-------|------|
 | Run F3-2 / F3-5 / F3-6 against Roberto's real fleet | Roberto | v0.4 milestone |
 | Land `cvg fleet rot`, `cvg fleet doc-drift`, `cvg fleet validate` CLI verbs | next sweep | unblocked by convergio-cli split |
-| ADR for extracting `fleet_*` routes from `convergio-server` | F4 | before next crate cap bump |
+| ~~ADR for extracting `fleet_*` routes from `convergio-server`~~ | ~~F4~~ | **landed (2026-05-19)**: pure mechanical split — no separate ADR needed. New crates `convergio-fleet-routes` + `convergio-server-core` (the latter holds the shared `AppState` + `ApiError`). `convergio-server` dropped from 14150 → 13075 lines; the 14000 → 14500 cap bump from F3-5 was rolled back. |
 | 5-repo daily-rebuild < 5 min measurement | observability | v0.4 milestone |
 
 ## Links
